move new tests on test_persistence file, rework teh new tests
